[Style]: style.ts 파일 생성 & 댓글, 공유 이미지 추가
- style 파일을 추가했습니다.
- 댓글, 북마크 이미지와 댓글 수, 북마크 수를 추가합니다.
